Do not depend strictly on TSDB connection, but on a TSDB adapter.
To do that, we've created a 'Historian', which is a metric collector. There's a historian for CloudWatch and for TSDB, for now,
and CloudWatch is still to be implemented.

This commit should be seamless and does not change behavior, but now we have a way to plug in different types of metric
collectors.

[#45654133]